requirements.txt
Django
psycopg2
SQLAlchemy
djangorestframework
pyjwt
